Fix(lint): Extract truncation logic to util and resolve title prop type conflict
